refactor batchnorm code, computation part extracted to normalize.xx
